Cleaning reset flow with hooks
